review updates: create AEPPushTemplate and extended types
- also add missing layout files
