+++ code/campaignclassic/src/main/res/layout/push_template_collapsed.xml
@@ -1,0 +1,33 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/basic_small_layout"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:orientation="vertical">
+
+    <TextView
+        android:id="@+id/notification_title"
+        style="@style/TextAppearance.Compat.Notification.Title"
+        android:layout_width="wrap_content"
+        android:layout_height="20dp"
+        android:layout_gravity="start"
+        android:textAlignment="viewStart"
+        android:text="small_notification_title_string" />
+
+    <TextView
+        android:id="@+id/notification_body"
+        style="@style/TextAppearance.Compat.Notification.Line2"
+        android:layout_width="wrap_content"
+        android:layout_height="30dp"
+        android:layout_gravity="start"
+        android:textAlignment="viewStart"
+        android:layout_below="@id/notification_title"
+        android:text="small_notification_body_string" />
+
+    <ImageView
+        android:id="@+id/template_image"
+        android:layout_width="match_parent"
+        android:layout_height="50dp"
+        android:layout_gravity="center"
+        android:layout_toRightOf="@id/notification_body" />
+</RelativeLayout>
